# back/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PatientStudiesResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/PatientStudyOut"
            },
            "title": "Items",
            "type": "array"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total"
        ],
        "title": "PatientStudiesResponse",
        "type": "object"
      },
...
      "StudyClinicalReportVersionsOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StudyClinicalReportVersionOut"
            },
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "StudyClinicalReportVersionsOut",
        "type": "object"
      },
...
      "StudyListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StudyDetailOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "StudyListResponse",
        "type": "object"
      },
...
      "StudyStatus": {
        "enum": [
          "in_progress",
          "completed",
          "cancelled",
          "scheduled"
        ],
        "title": "StudyStatus",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "Holter ECG API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/patients/{patient_id}/studies": {
      "get": {
        "operationId": "get_patient_studies_patients__patient_id__studies_get",
        "parameters": [
          {
            "in": "path",
            "name": "patient_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Patient Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientStudiesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Patient Studies",
        "tags": [
          "patients"
        ]
      }
    },
...
    "/studies": {
      "get": {
        "operationId": "list_studies_studies_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/StudyStatus"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "status[]",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "items": {
                    "$ref": "#/components/schemas/StudyStatus"
                  },
                  "type": "array"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status[]"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Studies",
        "tags": [
          "studies"
        ]
      }
    },
    "/studies/{study_id}": {
      "get": {
        "operationId": "get_study_studies__study_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Study Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Study",
        "tags": [
          "studies"
        ]
      }
    },
    "/studies/{study_id}/cancel": {
      "post": {
        "description": "Descarta el estudio: colocación fallida, datos de banco, error de carga.",
        "operationId": "cancel_study_studies__study_id__cancel_post",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Study Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyDetailOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Cancel Study",
        "tags": [
          "studies"
        ]
      }
    },
    "/studies/{study_id}/clinical-report/draft": {
      "get": {
        "operationId": "get_clinical_report_draft_studies__study_id__clinical_report_draft_get",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Study Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyClinicalReportDraftOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Clinical Report Draft",
        "tags": [
          "studies"
        ]
      },
      "put": {
        "operationId": "update_clinical_report_draft_studies__study_id__clinical_report_draft_put",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Study Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StudyClinicalReportDraftUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyClinicalReportDraftOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Update Clinical Report Draft",
        "tags": [
          "studies"
        ]
      }
    },
    "/studies/{study_id}/clinical-report/finalize": {
      "post": {
        "operationId": "finalize_clinical_report_studies__study_id__clinical_report_finalize_post",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Study Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "draftRevision",
            "required": true,
            "schema": {
              "minimum": 1,
              "title": "Draftrevision",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "snapshotHash",
            "required": true,
            "schema": {
              "maxLength": 64,
              "minLength": 64,
              "title": "Snapshothash",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/pdf": {
              "schema": {
                "contentMediaType": "application/octet-stream",
                "title": "Pdf",
                "type": "string"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyClinicalReportVersionOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Finalize Clinical Report",
        "tags": [
          "studies"
        ]
      }
    },
    "/studies/{study_id}/clinical-report/preview": {
      "get": {
        "operationId": "get_clinical_report_preview_studies__study_id__clinical_report_preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Study Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyClinicalReportPreviewOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Clinical Report Preview",
        "tags": [
          "studies"
        ]
      }
    },
    "/studies/{study_id}/clinical-reports": {
      "get": {
        "operationId": "list_clinical_report_versions_studies__study_id__clinical_reports_get",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Study Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StudyClinicalReportVersionsOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Clinical Report Versions",
        "tags": [
          "studies"
        ]
      }
    },
    "/studies/{study_id}/clinical-reports/{report_id}/pdf": {
      "get": {
        "operationId": "download_clinical_report_studies__study_id__clinical_reports__report_id__pdf_get",
        "parameters": [
          {
            "in": "path",
            "name": "study_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Study Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "in": "cookie",
            "name": "holter_session_v2",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Holter Session V2"
            }
          },
          {
            "in": "cookie",
            "name": "session",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Download Clinical Report",
        "tags": [
          "studies"
        ]
      }
    },
...
  }
}